更新管理员统计页面

+++ demo/src/main/java/com/example/demo/dto/UserPurchaseDTO.java
@@ -7,5 +7,5 @@ import lombok.Data;
 @AllArgsConstructor
 public class UserPurchaseDTO {
     private String nickname;
-    private int totalAmount;
+    private Long totalAmount;
 }
